chore(backend): Upgrades from @graphl-yoga/node 2.13.13 to graphql-yoga 3.9.1
