+++ apps/backend/config/customMiddleware.js
@@ -1,4 +1,4 @@
-import { createRequestHandler } from '../api/graphql'
+import yoga from '../api/graphql'
 import bodyParser from 'body-parser'
 import kue from 'kue'
 import kueUI from 'kue-ui'
@@ -27,13 +27,13 @@ export default function (app) {
   app.use('/admin/kue/api', kue.app)
   app.use('/admin/kue', kueUI.app)
 
-  app.use('/noo/graphql', cors({
+  app.use(yoga.graphqlEndpoint, cors({
     origin: '*',
     methods: 'GET, POST, PUT, DELETE, OPTIONS, HEAD',
     credentials: true
   }))
 
-  app.use('/noo/graphql', accessTokenAuth) // TODO: remove once all our URLs use JWTs
-  app.use('/noo/graphql', checkClientCredentials) // To auth API calls
-  app.use('/noo/graphql', createRequestHandler())
+  app.use(yoga.graphqlEndpoint, accessTokenAuth) // TODO: remove once all our URLs use JWTs
+  app.use(yoga.graphqlEndpoint, checkClientCredentials) // To auth API calls
+  app.use(yoga.graphqlEndpoint, yoga)
 }
