+++ apps/backend/package.json
@@ -63,9 +63,7 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/register": "^7.17.7",
     "@emoji-mart/data": "^1.0.6",
-    "@envelop/core": "^2.6.0",
-    "@graphql-yoga/node": "^2.13.13",
-    "@graphql-yoga/redis-event-target": "^0.1.3",
+    "@graphql-yoga/redis-event-target": "^3.0.3",
     "@hylo/shared": "workspace:*",
     "@mapbox/mapbox-sdk": "^0.13.3",
     "@sailshq/connect-redis": "^6.1.3",
@@ -97,6 +95,7 @@
     "graphql": "^15.3.0",
     "graphql-tools": "^9.0.1",
     "graphql-type-json": "^0.3.2",
+    "graphql-yoga": "^3.9.1",
     "grunt": "1.3.0",
     "image-size": "^0.3.5",
     "insane": "^2.6.2",
